Remove postgresql
This was used by past Rails projects, but can "conflict" i.e. cause some unexpected behavior in other projects that use their own PostgreSQL.

Ultimately the latter approach is better, and it's a bit odd to have a specific project dependency installed globally, and running in the background all the time.