Re-enable Starship Character module
Now that Warp supports "same line prompt" the Starship config can be simplified and the default Character module re-enabled.

End result is the same setup in Warp i.e. the prompt is on its own line. The Starship Character module is now shown before the cursor on the line after the prompt.

https://starship.rs/config/#character
https://docs.warp.dev/appearance/prompt#same-line-prompt